requirements.txt
fvcore==0.1.5.post20221221
git+https://github.com/stevewongv/InstanceShadowDetection.git@202fa2909243cfe2bfd817eb51d248cf553e0413#subdirectory=PythonAPI
detectron2@git+https://github.com/facebookresearch/detectron2.git@d4a5f28e01b2babbaba9f90198fb95f5c661cccd
opencv-python==4.5.5.64
openai-whisper==20230314
gradio==3.28.1
transformers==4.28.0
diffusers==0.15.1
controlnet_aux
langchain==0.0.101
decord==0.6.0
openai==0.27.4
ffmpeg-python==0.2.0
fairscale==0.4.3
lvis==0.5.3
suno-bark
easyocr==1.6.2
easydict==1.9
scikit-learn==0.24.2
pytorch-lightning==1.5.10
albumentations==0.5.2
webdataset==0.2.48
kornia==0.5.0
sentencepiece==0.1.98
accelerate==0.18.0
timm==0.6.13
git+https://github.com/facebookresearch/segment-anything.git
wget
protobuf==3.20
pytorchvideo==0.1.5
soundfile
decord==0.6.0
ftfy
regex
einops
imageio-ffmpeg
mediapipe
ipdb
gdown
addict
yapf
pillow==8.4.0